Add New View Doctor UI
