+++ .gitignore
@@ -46,4 +46,5 @@ test-results/
 
 # Instructions
 instructions/
+CLAUDE.local.md
